simplify and move NodeWithSpec to test_util
